Moved length_of_iterator to numba.core.inline_closurecall
